Move webhook only test to its own module, clean up simulations folder
