+++ testsuites/simulations/README.md
@@ -1,10 +1,10 @@
-### Running the simulation (End to end, XATTRs and the travel sample)
+### Running the Travel.NET simulation (End to end, XATTRs and the travel sample)
 
 1. Download Couchbase Server (5.0+) / Sync Gateway (1.5+)
 2. Create travel sample bucket / bucket user on Couchbase Server
 3. Start Sync Gateway targeting Couchbase Server
 ```
-./sync_gateway testsuites/simulations/sync.json 2>&1 | tee sgoutput.log
+./sync_gateway testsuites/simulations/Travel.NET/sync.json 2>&1 | tee sgoutput.log
 ```
 4. Open Travel.NET/Travel.NET.sln
 5. Run the app
@@ -16,3 +16,7 @@
     - This will update all of the docs via the Couchbase Server Python SDK
 8. Once, the python script exits, return to the app output and press enter
     - The app will now poll until it sees all of the SDK revsions. 
+
+### Running Todo simulation
+
+Currently the Todo directory is a collection of scripts for interaction with a live https://github.com/couchbaselabs/mobile-training-todo setup
